move joblevel verification

+++ OpenNos.Handler/WorldPacketHandler.cs
@@ -496,42 +496,48 @@ namespace OpenNos.Handler
                         {
                             if (!(Session.Character.UseSp && iteminfo.EquipmentSlot == (short)EquipmentType.Sp))
                             {
-
-                                Inventory equip = Session.Character.EquipmentList.LoadBySlotAndType(iteminfo.EquipmentSlot, (short)InventoryType.Equipment);
-                                if (equip == null)
+                                if (Session.Character.JobLevel >= iteminfo.LevelJobMinimum)
                                 {
-                                    inventory.Type = (short)InventoryType.Equipment;
-                                    inventory.Slot = iteminfo.EquipmentSlot;
+                                    Inventory equip = Session.Character.EquipmentList.LoadBySlotAndType(iteminfo.EquipmentSlot, (short)InventoryType.Equipment);
+                                    if (equip == null)
+                                    {
+                                        inventory.Type = (short)InventoryType.Equipment;
+                                        inventory.Slot = iteminfo.EquipmentSlot;
 
-                                    Session.Character.EquipmentList.InsertOrUpdate(ref inventory);
-                                    DeleteItem(type, slot);
+                                        Session.Character.EquipmentList.InsertOrUpdate(ref inventory);
+                                        DeleteItem(type, slot);
 
-                                    Session.Client.SendPacket(Session.Character.GenerateStatChar());
-                                    Thread.Sleep(100);
-                                    ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateEq(), ReceiverType.AllOnMap);
-                                    Session.Client.SendPacket(Session.Character.GenerateEquipment());
-                                    ClientLinkManager.Instance.Broadcast(Session, Session.Character.GeneratePairy(), ReceiverType.AllOnMap);
+                                        Session.Client.SendPacket(Session.Character.GenerateStatChar());
+                                        Thread.Sleep(100);
+                                        ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateEq(), ReceiverType.AllOnMap);
+                                        Session.Client.SendPacket(Session.Character.GenerateEquipment());
+                                        ClientLinkManager.Instance.Broadcast(Session, Session.Character.GeneratePairy(), ReceiverType.AllOnMap);
+                                    }
+                                    else
+                                    {
+                                        inventory.Type = (short)InventoryType.Equipment;
+                                        inventory.Slot = iteminfo.EquipmentSlot;
+
+                                        equip.Slot = slot;
+                                        equip.Type = type;
+
+                                        DeleteItem(type, slot);
+                                        Session.Character.EquipmentList.DeleteFromSlotAndType(slot, type);
+
+                                        Session.Character.InventoryList.InsertOrUpdate(ref equip);
+                                        Session.Character.EquipmentList.InsertOrUpdate(ref inventory);
+                                        Session.Client.SendPacket(Session.Character.GenerateInventoryAdd(equip.InventoryItem.ItemVNum, equip.InventoryItem.Amount, type, equip.Slot, equip.InventoryItem.Rare, equip.InventoryItem.Color, equip.InventoryItem.Upgrade));
+
+                                        Session.Client.SendPacket(Session.Character.GenerateStatChar());
+                                        Thread.Sleep(100);
+                                        ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateEq(), ReceiverType.AllOnMap);
+                                        Session.Client.SendPacket(Session.Character.GenerateEquipment());
+                                        ClientLinkManager.Instance.Broadcast(Session, Session.Character.GeneratePairy(), ReceiverType.AllOnMap);
+                                    }
                                 }
                                 else
                                 {
-                                    inventory.Type = (short)InventoryType.Equipment;
-                                    inventory.Slot = iteminfo.EquipmentSlot;
-
-                                    equip.Slot = slot;
-                                    equip.Type = type;
-
-                                    DeleteItem(type, slot);
-                                    Session.Character.EquipmentList.DeleteFromSlotAndType(slot, type);
-
-                                    Session.Character.InventoryList.InsertOrUpdate(ref equip);
-                                    Session.Character.EquipmentList.InsertOrUpdate(ref inventory);
-                                    Session.Client.SendPacket(Session.Character.GenerateInventoryAdd(equip.InventoryItem.ItemVNum, equip.InventoryItem.Amount, type, equip.Slot, equip.InventoryItem.Rare, equip.InventoryItem.Color, equip.InventoryItem.Upgrade));
-
-                                    Session.Client.SendPacket(Session.Character.GenerateStatChar());
-                                    Thread.Sleep(100);
-                                    ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateEq(), ReceiverType.AllOnMap);
-                                    Session.Client.SendPacket(Session.Character.GenerateEquipment());
-                                    ClientLinkManager.Instance.Broadcast(Session, Session.Character.GeneratePairy(), ReceiverType.AllOnMap);
+                                    Session.Client.SendPacket(Session.Character.GenerateMsg(Language.Instance.GetMessageFromKey("LOW_JOB_LVL"), 0));
                                 }
                             }
                             else
@@ -809,10 +815,10 @@ namespace OpenNos.Handler
                 foreach (ShopItem item in shop.ShopItems.Where(s => s.Type.Equals(type)))
                 {
                     Item iteminfo = ServerManager.GetItem(item.ItemVNum);
-                    if(iteminfo.Type != 0)
+                    if (iteminfo.Type != 0)
                         shoplist += String.Format(" {0}.{1}.{2}.{3}.{4}", iteminfo.Type, item.Slot, item.ItemVNum, -1, ServerManager.GetItem(item.ItemVNum).Price);
                     else
-                        shoplist += String.Format(" {0}.{1}.{2}.{3}.{4}.{5}", iteminfo.Type, item.Slot, item.ItemVNum, item.Rare,iteminfo.Colored?item.Color : item.Upgrade, ServerManager.GetItem(item.ItemVNum).Price);
+                        shoplist += String.Format(" {0}.{1}.{2}.{3}.{4}.{5}", iteminfo.Type, item.Slot, item.ItemVNum, item.Rare, iteminfo.Colored ? item.Color : item.Upgrade, ServerManager.GetItem(item.ItemVNum).Price);
 
                 }
 
@@ -1858,57 +1864,51 @@ namespace OpenNos.Handler
             Inventory sp = Session.Character.EquipmentList.LoadBySlotAndType((short)EquipmentType.Sp, (short)InventoryType.Equipment);
             Inventory fairy = Session.Character.EquipmentList.LoadBySlotAndType((short)EquipmentType.Fairy, (short)InventoryType.Equipment);
 
-            if (sp != null && Session.Character.JobLevel >= ServerManager.GetItem(sp.InventoryItem.ItemVNum).LevelJobMinimum)
+            if (Session.Character.Reput >= ServerManager.GetItem(sp.InventoryItem.ItemVNum).ReputationMinimum)
             {
-                if (Session.Character.Reput >= ServerManager.GetItem(sp.InventoryItem.ItemVNum).ReputationMinimum)
+                if (!(fairy != null && ServerManager.GetItem(fairy.InventoryItem.ItemVNum).Element != ServerManager.GetItem(sp.InventoryItem.ItemVNum).Element))
                 {
-                    if (!(fairy != null && ServerManager.GetItem(fairy.InventoryItem.ItemVNum).Element != ServerManager.GetItem(sp.InventoryItem.ItemVNum).Element))
-                    {
 
 
-                        Session.Character.UseSp = true;
-                        Session.Character.Morph = ServerManager.GetItem(sp.InventoryItem.ItemVNum).Morph;
-                        Session.Character.MorphUpgrade = sp.InventoryItem.Upgrade;
-                        Session.Character.MorphUpgrade2 = sp.InventoryItem.Color;
-                        ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateCMode(), ReceiverType.AllOnMap);
+                    Session.Character.UseSp = true;
+                    Session.Character.Morph = ServerManager.GetItem(sp.InventoryItem.ItemVNum).Morph;
+                    Session.Character.MorphUpgrade = sp.InventoryItem.Upgrade;
+                    Session.Character.MorphUpgrade2 = sp.InventoryItem.Color;
+                    ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateCMode(), ReceiverType.AllOnMap);
 
 
-                        /*s = "ski 833 833 833 834 835 836 837 838 839 840 841 21 25 28 37 41 44 49 53 56 340 341 345 352";
-                        MainFile.maps.SendMap(chara, s, true);
-                        /*
-                         qslot 0 1.1.2 1.1.1 1.1.3 0.7.-1 1.1.0 0.7.-1 0.7.-1 0.1.10 1.3.2 1.3.1
+                    /*s = "ski 833 833 833 834 835 836 837 838 839 840 841 21 25 28 37 41 44 49 53 56 340 341 345 352";
+                    MainFile.maps.SendMap(chara, s, true);
+                    /*
+                     qslot 0 1.1.2 1.1.1 1.1.3 0.7.-1 1.1.0 0.7.-1 0.7.-1 0.1.10 1.3.2 1.3.1
 
-                         qslot 1 1.1.2 1.1.3 1.1.4 1.1.5 1.1.6 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1
+                     qslot 1 1.1.2 1.1.3 1.1.4 1.1.5 1.1.6 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1
 
-                         qslot 2 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1
-                         */
+                     qslot 2 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1 7.7.-1
+                     */
 
 
 
-                        //  lev 40 2288403 14 72745 3221180 145000 20086 5
+                    //  lev 40 2288403 14 72745 3221180 145000 20086 5
 
-                        ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateEff(196), ReceiverType.AllOnMap);
+                    ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateEff(196), ReceiverType.AllOnMap);
 
-                        ClientLinkManager.Instance.Broadcast(Session, String.Format("guri 6 1 {0} 0 0", Session.Character.CharacterId), ReceiverType.AllOnMap);
-                        Session.Client.SendPacket(Session.Character.GenerateSpPoint());
-                        Session.Character.Speed += ServerManager.GetItem(sp.InventoryItem.ItemVNum).Speed;
-                        ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateCond(), ReceiverType.AllOnMap);
-                        Session.Client.SendPacket(Session.Character.GenerateLev());
-                    }
-                    else
-                    {
-                        Session.Client.SendPacket(Session.Character.GenerateMsg(Language.Instance.GetMessageFromKey("BAD_FAIRY"), 0));
-                    }
+                    ClientLinkManager.Instance.Broadcast(Session, String.Format("guri 6 1 {0} 0 0", Session.Character.CharacterId), ReceiverType.AllOnMap);
+                    Session.Client.SendPacket(Session.Character.GenerateSpPoint());
+                    Session.Character.Speed += ServerManager.GetItem(sp.InventoryItem.ItemVNum).Speed;
+                    ClientLinkManager.Instance.Broadcast(Session, Session.Character.GenerateCond(), ReceiverType.AllOnMap);
+                    Session.Client.SendPacket(Session.Character.GenerateLev());
                 }
                 else
                 {
-                    Session.Client.SendPacket(Session.Character.GenerateMsg(Language.Instance.GetMessageFromKey("LOW_REPUT"), 0));
+                    Session.Client.SendPacket(Session.Character.GenerateMsg(Language.Instance.GetMessageFromKey("BAD_FAIRY"), 0));
                 }
             }
             else
             {
-                Session.Client.SendPacket(Session.Character.GenerateMsg(Language.Instance.GetMessageFromKey("LOW_JOB_LVL"), 0));
+                Session.Client.SendPacket(Session.Character.GenerateMsg(Language.Instance.GetMessageFromKey("LOW_REPUT"), 0));
             }
+
         }
         public void ChangeVehicule(Item item)
         {
